requirements.txt
asgiref==3.4.1
cloudinary==1.28.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-phonenumber-field==6.0.0
django-summernote==0.8.20.0
gunicorn==20.1.0
phonenumberslite==8.12.41
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
